fix: remove router + move keys file to utils
