+++ core/components/Header.vue
@@ -1,6 +1,6 @@
 <template>
   <v-app-bar color="primary">
-    <v-app-bar-title class="cursor-pointer">{{ title }}</v-app-bar-title>
+    <v-app-bar-title>{{ title }}</v-app-bar-title>
     <v-toolbar-items v-if="eodash.routes">
       <v-btn v-for="route in eodash.routes" :key="route.to" variant="text" @click="navigateTo(route.to)">
         {{ route.title }}
@@ -10,9 +10,8 @@
   </v-app-bar>
 </template>
 <script setup>
-import { eodashKey } from '@/store/Keys';
+import { eodashKey } from '@/utils/keys';
 import { inject } from 'vue';
-// import { useRouter } from 'vue-router';
 
 const eodash = /** @type {import("@/types").Eodash} */(inject(eodashKey))
 
@@ -24,11 +23,7 @@ const title = eodash.brand?.name
  * @param {string} to
  */
 const navigateTo = (to) => {
-  // if (to.toLowerCase().startsWith('http')) {
   window.open(to, '_self')
-  // } else {
-  //   // push(to)
-  // }
 }
 </script>
 <style scoped lang='scss'>
