+++ core/eodash.js
@@ -31,7 +31,7 @@ export const eodash = reactive({
       id: Symbol(),
       type: "web-component",
       widget: {
-        //https://uiball.com/ldrs/
+        // https://uiball.com/ldrs/
         link: "https://cdn.jsdelivr.net/npm/ldrs/dist/auto/mirage.js",
         tagName: "l-mirage",
         properties: {
@@ -64,9 +64,7 @@ export const eodash = reactive({
         id: Symbol(),
         type: "internal",
         title: "datepicker",
-        // layout: { x: 3, y: 8, w: 6, h: 5 },
         layout: { x: 5, y: 11, w: 2, h: 1 },
-        // slidable: true,
         slidable: false,
         widget: {
           name: "EodashDatePicker",
