+++ core/plugins/index.js
@@ -1,15 +1,7 @@
-/**
- * plugins/index.ts
- *
- * Automatically included in `./src/main.ts`
- */
-
-// Plugins
 import vuetify from './vuetify';
-// import router from './router';
 import { createPinia } from 'pinia';
 import eodash from '@/eodash';
-import { eodashKey } from '@/store/Keys';
+import { eodashKey } from '@/utils/keys';
 import store from '../store';
 import { createHead } from '@unhead/vue'
 
@@ -24,7 +16,6 @@ export function registerPlugins(app) {
 
   app.use(vuetify)
     .use(head)
-    // .use(router)
     .use(pinia)
     .provide(eodashKey, eodash);
 }
