--- eodash-example.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="./.eodash/dist/style.css">
-</head>
-
-<body>
-    <eo-dash />
-    <script type="module" src="./.eodash/dist/eo-dash.js"></script>
-</body>
-
-</html>
