+++ package.json
@@ -49,7 +49,6 @@
     "vite": "^5.1.5",
     "vite-plugin-vuetify": "^2.0.0",
     "vue": "^3.2.0",
-    "vue-router": "^4.0.0",
     "vuetify": "^3.5.1",
     "webfontloader": "^1.6.28"
   },
